[finelog] Hoist schema/policy to top level; decouple client from store
Move finelog.store.schema and finelog.store.policy (dependency leaves that
import only finelog.errors and finelog.rpc) to finelog.schema and
finelog.policy, and relocate LOG_REGISTERED_SCHEMA into finelog.schema.
finelog.client no longer imports finelog.store at all — importing the client
now loads zero store modules — which removes the schema-type coupling that
blocked deleting finelog.store once the Iris in-process test surface migrates.

Also consolidate the duplicated parity test helpers (_worker_schema,
_worker_arrow_schema, _wait_for_health) into the parity conftest as the single
source; test_cutover keeps its own 4-column "note" variant, which is not a
duplicate.
