+++ lib/finelog/src/finelog/schema.py
@@ -120,6 +120,22 @@ class Schema:
         return tuple(c.name for c in self.columns)
 
 
+# The user-declared schema for the "log" namespace. The registry stamps
+# the implicit ``seq`` column on top.
+LOG_REGISTERED_SCHEMA = Schema(
+    columns=(
+        Column(name="key", type=stats_pb2.COLUMN_TYPE_STRING, nullable=False),
+        Column(name="source", type=stats_pb2.COLUMN_TYPE_STRING, nullable=False),
+        Column(name="data", type=stats_pb2.COLUMN_TYPE_STRING, nullable=False),
+        Column(name="epoch_ms", type=stats_pb2.COLUMN_TYPE_INT64, nullable=False),
+        Column(name="level", type=stats_pb2.COLUMN_TYPE_INT32, nullable=False),
+    ),
+    # Per-source tail reads (``WHERE key = $key ORDER BY seq DESC``) dominate;
+    # sorting by ``key`` first colocates same-source rows for row-group pruning.
+    key_column="key",
+)
+
+
 # ---------------------------------------------------------------------------
 # Conversions: proto / Arrow / JSON.
 # ---------------------------------------------------------------------------
